+++ src/patch/patch.h
@@ -17,7 +17,7 @@ STRUCT(PATCH) {
 };
 
 PATCH* patch_find(lua_State* L, const char* fileName);
-void patch_apply(lua_State* L, const char* fileName, PATCH* patch, void** fileData, unsigned long long* fileSize);
+void patch_apply(lua_State* L, const char* fileName, PATCH* patch, void** fileData, size_t* fileSize);
 
 void Patch_InitLua(lua_State* L);
